requirements.txt
# Flask and core dependencies
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Werkzeug==2.3.7

# Database and caching
redis==5.0.1

# Security and encryption
bcrypt==4.0.1
cryptography==41.0.7
PyJWT==2.8.0

# Validation and serialization
marshmallow==3.20.1
email-validator==2.1.0
phonenumbers==8.13.19

# Date and time handling
python-dateutil==2.8.2

# Environment and configuration
python-dotenv==1.0.0

# HTTP requests and utilities
requests==2.31.0
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
Jinja2==3.1.2

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-mock==3.12.0
pytest-cov==4.1.0
flake8==6.0.0
black==23.7.0

# Production server
gunicorn==21.2.0

# Monitoring and logging
sentry-sdk[flask]==1.38.0
python-json-logger==2.0.7

# API Documentation
flasgger==0.9.7.1